--- a/docs/src/whatsnew/latest.rst
+++ b/docs/src/whatsnew/latest.rst
@@ -94,7 +94,7 @@ This document explains the changes made to Iris for this release
    :meth:`~iris.coords.DimCoord.from_regular`. (:pull:`4698`)
 
 #. `@pp-mo`_ streamlined :meth:`~iris.cube.Cube.add_aux_factory`.
-   (:pull:`xxxx`)
+   (:pull:`4718`)