In response to https://github.com/SciTools/iris/issues/4713 and several similar benchmark performance triggers, 
which seem to indicate that the time to call `iris.cube.Cube.add_aux_factory` is highly variable.
Actually, it seems on reflection that this time is unexpectedly **_long_**.

This attempts to address the speed issue, by 
Initial local testing results ...
```
$ nox --session="benchmarks(custom)" -- continuous -e --bench AuxFactory "2f9c8d0da~1..c60ff9a80"
nox > Running session benchmarks(custom)
  . . .
Running ASV...
nox > cd benchmarks
nox > asv machine --yes
  . . .
· Ignoring ASV setting(s): `python`. Benchmark environment management is delegated to third party script(s).
· Creating environments
· Discovering benchmarks
· Running 6 total benchmarks (2 commits * 1 environments * 3 benchmarks)
[  0.00%] · For scitools-iris commit 2f9c8d0d <latest> (round 1/2):
[  0.00%] ·· Building for conda-delegated............
[  0.00%] ·· Benchmarking conda-delegated
[  8.33%] ··· Running (cube.AuxFactory.time_add--)....
[ 25.00%] · For scitools-iris commit c60ff9a8 <addaux_speedup> (round 1/2):
[ 25.00%] ·· Building for conda-delegated......
[ 25.00%] ·· Benchmarking conda-delegated
[ 33.33%] ··· Running (cube.AuxFactory.time_add--)...
[ 50.00%] · For scitools-iris commit c60ff9a8 <addaux_speedup> (round 2/2):
[ 50.00%] ·· Benchmarking conda-delegated
[ 58.33%] ··· cube.AuxFactory.time_add                                                                                                                                               77.6±0.3μs
[ 66.67%] ··· cube.AuxFactory.time_create                                                                                                                                            14.0±0.2μs
[ 75.00%] ··· cube.AuxFactory.time_return                                                                                                                                            85.7±0.5ns
[ 75.00%] · For scitools-iris commit 2f9c8d0d <latest> (round 2/2):
[ 75.00%] ·· Building for conda-delegated...
[ 75.00%] ·· Benchmarking conda-delegated
[ 83.33%] ··· cube.AuxFactory.time_add                                                                                                                                              7.23±0.06ms
[ 91.67%] ··· cube.AuxFactory.time_create                                                                                                                                            22.3±0.1μs
[100.00%] ··· cube.AuxFactory.time_return                                                                                                                                              88.2±1ns
.       before           after         ratio
     [2f9c8d0d]       [c60ff9a8]
     <latest>         <addaux_speedup>
-      22.3±0.1μs       14.0±0.2μs     0.63  cube.AuxFactory.time_create
-     7.23±0.06ms       77.6±0.3μs     0.01  cube.AuxFactory.time_add

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
nox > Session benchmarks(custom) was successful.

$ git status
```

If that's replicated, and tests all pass, it sounds like a win !