@trexfeathers 2022-04-26
```
-     7.23±0.06ms       77.6±0.3μs     0.01  cube.AuxFactory.time_add
```

Dude!